add transaction and wallet

+++ db/migrate/20230819205354_add_minimum_investment_column_to_asset.rb
@@ -1,6 +1,6 @@
 class AddMinimumInvestmentColumnToAsset < ActiveRecord::Migration[7.0]
   def change
-    add_column :assets, :minimum_investment, :float
+    add_column :assets, :minimum_investment, :decimal
     add_column :assets, :term, :string
     add_column :assets, :payment_schedule, :integer, default: 0
   end
