+++ db/migrate/20230823192547_add_amount_to_portfolio.rb
@@ -1,6 +1,6 @@
 class AddAmountToPortfolio < ActiveRecord::Migration[7.0]
   def change
-    add_column :portfolios, :amount, :float
+    add_column :portfolios, :amount, :decimal
     
   end
 end
